+++ tools/tesseratos/src/tesseratos/menu_bar/plugin.hpp
@@ -1,0 +1,22 @@
+/// @dir
+/// @brief @ref tesseratos-menu-bar-plugin plugin directory.
+
+/// @file
+/// @brief Plugin entry point.
+/// @ingroup tesseratos-menu-bar-plugin
+
+#pragma once
+
+#include <cubos/engine/prelude.hpp>
+
+namespace tesseratos
+{
+    /// @defgroup tesseratos-menu-bar-plugin Menu bar
+    /// @ingroup tesseratos
+    /// @brief Adds a menu bar to tesseratos.
+
+    /// @brief Plugin entry function.
+    /// @param cubos @b Cubos main class
+    /// @ingroup tesseratos-menu-bar-plugin
+    void menuBarPlugin(cubos::engine::Cubos& cubos);
+} // namespace tesseratos